Catch order error, creates user

+++ src/main/java/com/zaico/cms/servlets/user/UserDeleteServlet.java
@@ -4,6 +4,8 @@ import com.zaico.cms.entities.User;
 import com.zaico.cms.servicies.implementation.FactoryService;
 import com.zaico.cms.servicies.implementation.UserServiceImpl;
 import com.zaico.cms.servicies.interfaces.UserService;
+import com.zaico.cms.utility.ErrorCode;
+import com.zaico.cms.utility.ExceptionCMS;
 import com.zaico.cms.utility.ExceptionHandler;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -11,6 +13,7 @@ import org.apache.log4j.Logger;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -47,9 +50,15 @@ public class UserDeleteServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         try {
-            if ( request.getParameter("id") != null) {
-                Integer id = Integer.parseInt(request.getParameter("id"));
-                user = userService.findUser((long) id);
+
+            String sessionUser = "";
+            for (Cookie cookie : request.getCookies()) {
+                if (cookie.getName().equals("user")) {
+                    sessionUser = cookie.getValue();                }
+            }
+
+            if ( sessionUser == user.getLogin()) {
+                throw new ExceptionCMS("You can`t delete youself!", ErrorCode.USER_CANNOT_BE_DELETED);
             }
             userService.deleteUser(user);
             String message = "User \""+user.getLogin()+"\" deleted successfully";
